Fix viewport and scrollbars for top-controls hiding.
1. Simplify top controls manager to never be in "overlay" mode, i.e. the
top controls always push down the contents (note that the browser
compositor is responsible for this; the renderer compositor just draws
at the top of its surface and sends up the offset as metadata).

2. Introduce the concept of "visibleViewportSize".  This is the true
user-visible size on the current frame in DIP, not including
top-controls or non-overlay scrollbars.  This is used for max scroll bounds
and scrollbar positioning.

3. Just disable clip layer masksToBounds if top controls are present,
since it provides no value in the Android case.

4. Given the dynamism of the viewport size, abandon the WebKit scrollbar
code and manually compute scrollbar quads after all.  There are three
inputs: trackRect (which is the layer size plus some margins);
m_verticalAdjust which compensates for the fact the scrollbar layers are
positioned by WebKit; and m_normalizedViewport which is a rect that
encapsulates the position and size of both scrollbars.

NOTRY=true
BUG=161303

Review URL: https://chromiumcodereview.appspot.com/12519011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188445 0039d316-1c4b-4281-b951-d872f2087c98