cc: Set update rect properly after initializing tiles.
This change is focused on correctness and currently just sets the
update rect to the layer bounds. This makes impl-side painting work
correctly when using partial updates but doesn't yet take advantage
of it.

Next step is to only add the tile region to the update rect and take
full advantage of partial updates. Some re-factoring is required to
make that efficient.

This also fixes a problem with low resolution and solid color tile
initialization. Low resolution tiles were previously not triggering
a redraw. The result was that we sometimes showed checkerboarding
when we could show low resolution content.

BUG=170241


Review URL: https://chromiumcodereview.appspot.com/12848007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188612 0039d316-1c4b-4281-b951-d872f2087c98