Android: Work around driver bug reporting GL errors.
The driver on the S3 w/ARM Mali erroneously reports
GL_INVALID_FRAMEBUFFER_OPERATION when you call glGetError() while the
bound FBO is incomplete. This causes various DCHECKs to fail.

Simply unbind the FBO for the --composite-to-mailbox path when we drop
the backing or send the frame to the browser. This will cause the
default FBO to be bound by the command decoder.

BUG=196635
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12529021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188831 0039d316-1c4b-4281-b951-d872f2087c98