[Telemetry] Fix GetChildPids() on windows.
This fixes the process metrics on the windows page cyclers. The bug was that
wmic process doesn't output fields in the same order as they are specified. So
even though we specified ProcessId,ParentProcessId, it was outputting parent
first. So I corrected pid_ppid -> ppid, pid. I also changed the order we
specify the fields to enhance readability, but that is a functional no-op.

BUG=196411
TEST=python tools\perf\run_multipage_benchmarks --browser=system page_cycler tools\perf\page_sets\page_cycler\moz.json
NOTRY=True


Review URL: https://chromiumcodereview.appspot.com/12659006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188890 0039d316-1c4b-4281-b951-d872f2087c98