Remove MessageLoopRunner from ProfileSigninConfirmationDialogTest.
The current test uses a MessageLoopRunner to wait for some callbacks to
be invoked.  These callback own a reference to the MessageLoopRunner,
but Sometimes one of these callbacks is deleted on the history thread.

This change replaces the use of MessageLoopRunner with a stack-allocated
RunLoop instance, so that the creating thread is its only owner.

This change should fix an occasionally-detected error on the Chromium
Memory FYI Linux Tests (tsan) bots.

BUG=196513
TBR=jhawkins


Review URL: https://chromiumcodereview.appspot.com/12678006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188973 0039d316-1c4b-4281-b951-d872f2087c98