Switch xz from handwritten 32-bit x86 assembly code for CRC32 to a small portable C implementation and take further measures to reduce code size.
This allows xz to be built for all architectures, including x86_64. The C
version is even smaller in terms of code size than the 32-bit x86 assembly
version. These changes reduce the decompressor (which is the only portion that
ships, and it only ships with diff updaters) by a further 16kB, 10kB
compressed. Time-wise, for 32-bit x86, the "small" variant now in use results
in 8% slower decompression, but the times are so short already that nobody's
going to notice: the difference in absolute time spent decompressing on my
laptop was 4 milliseconds, and with updates normally applied in the
background, this is effectively a productivity no-op.

BUG=222086
Review URL: https://codereview.chromium.org/12939006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189100 0039d316-1c4b-4281-b951-d872f2087c98