On Windows 8 ASH we should not display the app list if the user has not signed in.
This does not affect Chrome OS. Currently on Windows 8 ASH if we click on the app launcher when
we are not signed in, you see a view which displays the sign in window on the left and the apps to its
right. Clicking on the apps does not do anything as we need to be signed in to chrome to be able to use
them.

Fix is to not display the apps if we need to be signed in.

BUG=180052
R=benwells
Review URL: https://codereview.chromium.org/12939008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189120 0039d316-1c4b-4281-b951-d872f2087c98