Fix gl_tests for Galaxy Nexus
Query bugs were real. Added more unit tests

CopyTextureCHROMIUM bugs are iffy. We could try to work around
them in the command buffer. Basically we'd have to set filtering
on all textures used in an FBO just before each glClear/glDraw/glRead

That would be slow so we could do that only on SGX. We could optimize
if we knew either (a) there are no textures with problematic filtering
(unlikely to be true) or (b) there are no textures with problemative filtering 
on the currently bound FBO.

It's not a small change so we can do that later if we want.

There's still an issue with mailboxes that I'll fix in another CL

BUG=180432


Review URL: https://chromiumcodereview.appspot.com/12716027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189291 0039d316-1c4b-4281-b951-d872f2087c98