Fixed an issue with the notification center layout.
A bug in MessageCenterView's Layout() was causing the notification
center buttons to be placed in the wrong location when notifications
were received while the notification center was updated.

Adding an additional layout step in Layout() is a brutish but effective
way to fix the problem. A more subtle fix can come later.

BUG=222221
R=miket@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12512012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189413 0039d316-1c4b-4281-b951-d872f2087c98