Removed skia::ScopedPlatformCanvas.
No changes in behavior. This is a trivial class that has only one actual use outside
of unit tests. 

The reason for the removal is specific to Windows component builds and somewhat
complicated. The pattern of exporting a class that inherits a non exported template
causes the template to be implicitly exported on Windows. That might cause multiple
definition linker errors for users of the template in other DLLs. For more info refer
to this blog post:
http://www.codesynthesis.com/~boris/blog/2010/01/18/dll-export-cxx-templates/

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12972002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189633 0039d316-1c4b-4281-b951-d872f2087c98