Generalzie GpuBlacklist to GpuControlList.
This CL refactor the GpuBlacklist so it can be configured to handle a different set of features.

So it can handle not only blacklist, but dual gpu switching, and driver bugs.

I will add the code to truly manage driver bug workarounds in the next CL.

BUG=222857
TEST=content_unittests
TBR=asargent
Review URL: https://codereview.chromium.org/12976004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@189889 0039d316-1c4b-4281-b951-d872f2087c98