Split apk-build.xml into 3 steps
In preparation for rewriting these actions in python, split apk-build.xml into smaller ant scripts. This should not change the functionality and mostly is just splitting the file and identifying the properties and defines required for each of the new files.

Some of the configuration is moved out of ant and into gyp (i.e. the android_manifest_path is now gyp-configurable).

Also, the ant scripts now use the correct basedir! That is, these ant scripts actually understand relative paths. To go along with this, resource_dir is the relative path to the resource directory (rather than a path relative to java_in_dir).

BUG=158821
TBR=torne@chromium.org, brettw@chromium.org
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12963009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@190712 0039d316-1c4b-4281-b951-d872f2087c98