[Telemetry] Prevent one form of possible infinite recursion in GetChildPids.
The win bots are occassionally infinitely recursing on GetChildPids (see bug).
I can't reproduce it locally, but this is one way I can think of it getting
stuck.

It should never be the case that one of the processes we are interested in is
its own parent, but just in case I'm adding this check.

If this doesn't fix it, the next step will be to add some logging.

BUG=223108
TEST=moz page cycler on windows
NOTRY=True

Review URL: https://codereview.chromium.org/13004025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191039 0039d316-1c4b-4281-b951-d872f2087c98