CrOS login: Fix handling of allow-new-users vs. user_whitelist.
It turns out that we have existing deployments that have an empty
user_whitelist and allow_new_users not set. The recent change to treat
that as the signal that sign in is verboten entirely broke them, so
revert that change and instead only prevent login entirely if
allow_new_users is explicitly set to off.

BUG=chromium:223161
TEST=unit tests


Review URL: https://chromiumcodereview.appspot.com/13183002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191273 0039d316-1c4b-4281-b951-d872f2087c98