Instant: Don't use the local overlay's URL for the NTP.
GetInstantURL() is used for both the NTP as well as the overlay. So, it's wrong
for it to assume that it's okay to return the local overlay's URL under some
circumstances. Instead, it should return false, and have the caller fallback to
the local NTP's URL or local overlay's URL as appropriate.

BUG=224946
R=samarth@chromium.org
TEST=See bug.

Review URL: https://chromiumcodereview.appspot.com/13162003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191967 0039d316-1c4b-4281-b951-d872f2087c98