Add new methods to DriveServiceInterface. These will replace GetResourceList.
This is the preparation to split the method GetResourceList, which has now
become complicated, and difficult to support all the functions on Drive API v2.
The main goal is, by splitting GetResourceList based on the requirement and
actual usage, supporting them on Drive API v2. This is the first step of the
work.

BUG=160932
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/13602003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@192289 0039d316-1c4b-4281-b951-d872f2087c98