History: Prevent entries from taking up all the available space.
Before the link entry would extend as much as it could to the right, this patch
prevents that and restores the previous behavior.

This is done by moving the div that contains the managed mode indicators (which
are not shown normally anyways) one level up in the DOM and making that div
take up available the space.

R=dubroy@chromium.org
BUG=224081


Review URL: https://chromiumcodereview.appspot.com/13534002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@192295 0039d316-1c4b-4281-b951-d872f2087c98