Adding WebRTC performance expectations.
This will be used to keep track of WebRTC's performance tests. These used to be in the internal build bot code for WebRTC, but we need to move them due to a recent chrome infra change. We can't put them in the regular tools/perf_expectations yet since we are not done migrating to chrome infra. This is expected to be temporary.

BUG=
NOTRY=True


Review URL: https://chromiumcodereview.appspot.com/13773008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@192820 0039d316-1c4b-4281-b951-d872f2087c98