cc: Prevent checkerboarded animations from timing out during tests.
During tests that verify behaviour with checkerboarding, we need
preductable behaviour. Currently if the test takes too long and
the compositor thread ticks too many times, the checkerboarded
animation will time out and animate+draw. This is unpredictable
flaky behaviour for the tests. So add a setting to disable this
timeout-and-draw behaviour in the scheduler.

R=jamesr
BUG=228998

Review URL: https://chromiumcodereview.appspot.com/13577005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193006 0039d316-1c4b-4281-b951-d872f2087c98