gpu: Disable virtual contexts for Qualcomm on SDK-16
We needed virtual contexts for Qualcomm, but we now
have a startup crasher after enabling composite-to-mailbox.
This disables virtual context on SDK-16 for beta until
we can find a better solution.

BUG=187487
NOTRY=true

No-try as this is android only and we need it for beta.

Review URL: https://chromiumcodereview.appspot.com/13913021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193207 0039d316-1c4b-4281-b951-d872f2087c98